+++ mobile/src/screens/LoadingScreen.tsx
@@ -1,4 +1,4 @@
-// import { useNavigation } from "@react-navigation/native";
+import { useNavigation } from "@react-navigation/native";
 import React, { useContext } from "react";
 import { Text, View, Image } from "react-native";
 
@@ -6,17 +6,28 @@ import Button from "../components/Button";
 import { UserContext } from "../contexts/UserContext";
 import { AppStylesheet } from "../styles/AppStylesheet";
 import stemettImage from "../../assets/Stemett.png";
+import { SocketContext } from "../contexts/SocketContext";
+
 
 const LoadingScreen = () => {
     // basically same code as HomeScreen.tsx
-    // const navigation = useNavigation();
+    const navigation = useNavigation();
     const { userState } = useContext(UserContext);
+
+    const socket = useContext(SocketContext);
+
+    const connectToGame = () => {
+      socket.emit("assign to room");
+      socket.once("successful assign", () => {
+        navigation.navigate("Chess");
+      });
+    };
 
     return (
         <View style={AppStylesheet.container}>
           <Text style={AppStylesheet.headerHomeScreen}>Welcome, {userState.firstName}</Text>
           <Image style={AppStylesheet.mascot} source={stemettImage} />
-          <Button text="Waiting for Opponent"/>
+          <Button text="Waiting for Opponent" onPress={connectToGame}/>
           <Text>14 Players Online</Text>
           <Button text = "Stop Searching"/>
         </View>
